requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
openai>=1.14.0
python-dotenv==1.0.0
requests==2.31.0
